create mno account, get mno account by id and remove mno account

+++ src/main/java/com/atn/mnomanager/logic/impl/MnoAccountProcessor.java
@@ -1,11 +1,11 @@
 package com.atn.mnomanager.logic.impl;
 
 import com.atn.mnomanager.entities.*;
+import com.atn.mnomanager.exceptions.HandlerInternalServerErrorException;
+import com.atn.mnomanager.exceptions.HandlerNotFoundException;
 import com.atn.mnomanager.facades.*;
 import com.atn.mnomanager.logic.IMnoAccountProcessor;
-import com.atn.mnomanager.logic.IMnoProductProcessor;
 import com.atn.mnomanager.models.MnoAccountModel;
-import com.atn.mnomanager.models.MnoProductModel;
 import com.atn.mnomanager.models.SuccessResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -23,7 +23,7 @@ public class MnoAccountProcessor implements IMnoAccountProcessor {
     private MnoProfileRepository mnoProfileRepository;
 
     @Autowired
-    private AccountRepository accoutRepository;
+    private AccountRepository accountRepository;
 
     /**
      * Create MnoAccount processor
@@ -34,11 +34,17 @@ public class MnoAccountProcessor implements IMnoAccountProcessor {
     @Override
     public MnoAccount createMnoAccount(MnoAccountModel mnoAccountModel) {
         try {
-            MnoProfile foundMnoProfile = mnoProfileRepository.findById(mnoAccountModel.getMnoProfileId()).get();
-            Account foundAccount = accoutRepository.findById(mnoAccountModel.getAccountId()).get();
+            MnoProfile foundMnoProfile = mnoProfileRepository.findById(mnoAccountModel.getMnoProfileId()).orElse(new MnoProfile());
+            if (!foundMnoProfile.getId().equals(mnoAccountModel.getMnoProfileId())) {
+                throw new HandlerNotFoundException("MNO not found");
+            }
+            Account foundAccount = accountRepository.findById(mnoAccountModel.getAccountId()).orElse(new Account());
+            if (!foundAccount.getId().equals(mnoAccountModel.getAccountId())) {
+                throw new HandlerNotFoundException("Account not found");
+            }
             return mnoAccountRepository.save(new MnoAccount(mnoAccountModel.getIsNormalAccount(), foundMnoProfile, foundAccount));
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            throw new HandlerInternalServerErrorException("Internal server error");
         }
     }
 
@@ -51,9 +57,13 @@ public class MnoAccountProcessor implements IMnoAccountProcessor {
     @Override
     public MnoAccount getAccountById(String id) {
         try {
-            return mnoAccountRepository.findById(id).get();
+            MnoAccount foundMnoAccount = mnoAccountRepository.findById(id).orElse(new MnoAccount());
+            if (!foundMnoAccount.getId().equals(id)) {
+                throw new HandlerNotFoundException("MnoAccount not found");
+            }
+            return foundMnoAccount;
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            throw new HandlerInternalServerErrorException("Internal server error");
         }
     }
 
@@ -66,10 +76,14 @@ public class MnoAccountProcessor implements IMnoAccountProcessor {
     @Override
     public SuccessResponse removeMnoAccount(String id) {
         try {
+            MnoAccount foundMnoAccount = mnoAccountRepository.findById(id).orElse(new MnoAccount());
+            if (!foundMnoAccount.getId().equals(id)) {
+                throw new HandlerNotFoundException("MnoAccount not found");
+            }
             mnoAccountRepository.deleteById(id);
             return new SuccessResponse("mno account well removed");
         } catch (Exception e) {
-            throw new RuntimeException(e);
+            throw new HandlerInternalServerErrorException("Internal server error");
         }
     }
 }
